requirements.txt
fastapi
uvicorn
sqlalchemy 
pymysql
asyncio
pydantic-settings
python-multipart
cryptography
python-jose
requests
firebase-admin
secure-smtplib
librosa 
scikit-learn 
numpy==1.21.6
python-multipart 
joblib
aiohttp
pandas
nltk
twilio
unidecode
websockets